requirements.txt
google-generativeai
requests
beautifulsoup4
streamlit
pyperclip
python-dotenv
